invenotry-export-from-store/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
python-dotenv==1.0.1
requests==2.32.3
